pyproject.toml
[project]
name = "agent-honk"
version = "0.1.0"
description = "Discord bot wrapper around Goose AI - Codename: Agent Honk"
authors = [{name = "Discord Goose Team", email = "team@example.com"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "F", "W", "I"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "ruff>=0.12.1",
]
